requirements.txt
fastapi
uvicorn
motor
bcrypt
PyJWT
python-dotenv
pydantic
